Create gt7-extramenus-original-optimized-slow-network.py
This is the original (playstation remote play, NOT Chiaki) script with alternate timing support removed, with longer timings tested for slower connections for remote play.  It will still bug out if your remote play FPS gets extremely choppy, which happens a lot for some reason, but the timings are much more tolerable with less errors on medium to stable connections.

To make it work with Chiaki (default keyboard mappings), replace all instances of 'escape' with 'backspace'.